Danki_with_Django/requirements.txt
aiohttp==3.7.3
pandas==1.1.4
Django==3.1.4
model_bakery==1.2.1
requests==2.25.0
aiofiles==0.6.0
genanki==0.9.1
djangorestframework==3.12.2
